(lib) move state includes internally
The externally-visible API needs not know about the internal structure
of the Casemate model state.

This moves the state typedefs out of casemate.h and into casemate-impl.h
and makes the initialisation function error out if not supplied enough space.

At the same time, drop to the standard (u)intN_s in the API header.